backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
paho-mqtt==2.0.0
influxdb-client==1.43.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
